changed pet weight from int to float

+++ src/main/java/com/teksystems/Osterbur_Erika_PetMed_CaseStudy/database/entity/VetVisit.java
@@ -35,7 +35,7 @@ public class VetVisit {
     private String notes;
 
     @Column(name = "weight")
-    private Integer weight;
+    private Float weight;
 
     @ToString.Exclude
     @ManyToOne
